# docs/swagger.yaml
basePath: /
definitions:
  errors.APIError:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
      status:
        type: integer
    type: object
...
  handlers.createPostRequest:
    properties:
      is_idea:
        type: boolean
      scheduled_at:
        type: string
      text_content:
        type: string
      title:
        type: string
      type:
        type: string
    type: object
...
  post.Post:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      id:
        type: string
      is_idea:
        type: boolean
      project_id:
        type: string
      scheduled_at:
        type: string
      status:
        type: string
      text_content:
        type: string
      title:
        type: string
      type:
        $ref: '#/definitions/post.PostType'
      updated_at:
        type: string
    type: object
  post.PostResponse:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      id:
        type: string
      is_idea:
        type: boolean
      linked_platforms:
        items:
          $ref: '#/definitions/post.Platform'
        type: array
      project_id:
        type: string
      scheduled_at:
        type: string
      status:
        type: string
      text_content:
        type: string
      title:
        type: string
      type:
        $ref: '#/definitions/post.PostType'
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API server for OpenCM
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: OpenCM API
  version: "1.0"
paths:
...
  /posts:
    get:
      consumes:
      - application/json
      description: Get available post types
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.APIError'
      security:
      - ApiKeyAuth: []
      summary: Get available post types
      tags:
      - posts
  /posts/{project_id}:
    get:
      consumes:
      - application/json
      description: List all posts of a project by its id
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/post.Post'
            type: array
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/errors.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.APIError'
        "410":
          description: Project not found
          schema:
            $ref: '#/definitions/errors.APIError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.APIError'
      security:
      - ApiKeyAuth: []
      summary: List all posts of a project
      tags:
      - posts
  /posts/{project_id}/{post_id}:
    delete:
      consumes:
      - application/json
      description: Delete a post by its id. We might or might not want to implement
        pagination and filtering. For the time being, we will keep it simple.
      parameters:
      - description: Post ID
        in: path
        name: post_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No content
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/errors.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.APIError'
        "410":
          description: Post not found
          schema:
            $ref: '#/definitions/errors.APIError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.APIError'
      security:
      - ApiKeyAuth: []
      summary: Delete a post
      tags:
      - posts
    get:
      consumes:
      - application/json
      description: Get a post by its id
      parameters:
      - description: Post ID
        in: path
        name: post_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/post.PostResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/errors.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.APIError'
        "410":
          description: Post not found
          schema:
            $ref: '#/definitions/errors.APIError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.APIError'
      security:
      - ApiKeyAuth: []
      summary: Get a post by id
      tags:
      - posts
    patch:
      consumes:
      - application/json
      description: Update a post with the given title, text content, image links,
        video links, is idea and scheduled at
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      - description: Post ID
        in: path
        name: post_id
        required: true
        type: string
      - description: Post update request
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/handlers.createPostRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No content
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/errors.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.APIError'
        "410":
          description: Post not found
          schema:
            $ref: '#/definitions/errors.APIError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.APIError'
      security:
      - ApiKeyAuth: []
      summary: Update a post
      tags:
      - posts
  /posts/{project_id}/{post_id}/archive:
    patch:
      consumes:
      - application/json
      description: Archive a post by its id
      parameters:
      - description: Post ID
        in: path
        name: post_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No content
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/errors.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.APIError'
        "410":
          description: Post not found
          schema:
            $ref: '#/definitions/errors.APIError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.APIError'
      security:
      - ApiKeyAuth: []
      summary: Archive a post
      tags:
      - posts
  /posts/{project_id}/{post_id}/dequeue:
    patch:
      consumes:
      - application/json
      description: Remove a post from a project queue by its id
      parameters:
      - description: Post ID
        in: path
        name: post_id
        required: true
        type: string
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No content
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/errors.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.APIError'
        "410":
          description: Post not found
          schema:
            $ref: '#/definitions/errors.APIError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.APIError'
      security:
      - ApiKeyAuth: []
      summary: Remove a post from a project queue
      tags:
      - posts
  /posts/{project_id}/{post_id}/enqueue:
    patch:
      consumes:
      - application/json
      description: Add a post to a project queue by its id
      parameters:
      - description: Post ID
        in: path
        name: post_id
        required: true
        type: string
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No content
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/errors.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.APIError'
        "410":
          description: Post not found
          schema:
            $ref: '#/definitions/errors.APIError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.APIError'
      security:
      - ApiKeyAuth: []
      summary: Add a post to a project queue
      tags:
      - posts
  /posts/{project_id}/{post_id}/platforms/{platform_id}:
    delete:
      consumes:
      - application/json
      description: Remove a social media publisher platform from a post by its id
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      - description: Post ID
        in: path
        name: post_id
        required: true
        type: string
      - description: Platform ID
        in: path
        name: platform_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No content
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/errors.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.APIError'
        "410":
          description: Post not found
          schema:
            $ref: '#/definitions/errors.APIError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.APIError'
      security:
      - ApiKeyAuth: []
      summary: Remove a social media publisher platform from a post
      tags:
      - posts
    post:
      consumes:
      - application/json
      description: Add a social media publisher platform to a post by its id
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      - description: Post ID
        in: path
        name: post_id
        required: true
        type: string
      - description: Platform ID
        in: path
        name: platform_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No content
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/errors.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.APIError'
        "410":
          description: Post not found
          schema:
            $ref: '#/definitions/errors.APIError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.APIError'
      security:
      - ApiKeyAuth: []
      summary: Add a social media publisher platform to a post
      tags:
      - posts
  /posts/{project_id}/{post_id}/restore:
    patch:
      consumes:
      - application/json
      description: Restore a post by its id
      parameters:
      - description: Post ID
        in: path
        name: post_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No content
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/errors.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.APIError'
        "410":
          description: Post not found
          schema:
            $ref: '#/definitions/errors.APIError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.APIError'
      security:
      - ApiKeyAuth: []
      summary: Restore a post
      tags:
      - posts
  /posts/{project_id}/{post_id}/schedule:
    patch:
      consumes:
      - application/json
      description: Schedule a post by its id
      parameters:
      - description: Post ID
        in: path
        name: post_id
        required: true
        type: string
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      - description: Scheduled at
        in: body
        name: scheduled_at
        required: true
        schema:
          $ref: '#/definitions/handlers.schedulePostRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No content
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/errors.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.APIError'
        "410":
          description: Post not found
          schema:
            $ref: '#/definitions/errors.APIError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.APIError'
      security:
      - ApiKeyAuth: []
      summary: Schedule a post
      tags:
      - posts
...
schemes:
- http
- https
swagger: "2.0"